requirements.txt
# Deep Learning Framework
torch==2.1.0
torchvision==0.16.0
timm==0.9.12
albumentations==1.3.1

# Data Processing
numpy==1.24.3
pandas==2.0.3
opencv-python==4.8.0.74
scikit-learn==1.3.0

# Metrics and Visualization
wandb==0.15.12
matplotlib==3.7.2
seaborn==0.12.2

# Utilities
tqdm==4.66.1
PyYAML==6.0.1
hydra-core==1.3.2
omegaconf==2.3.0

# Development
black==23.7.0
isort==5.12.0
pylint==2.17.5
pytest==7.4.0
